fix(cockpit): close isomorphic-route server-only leak into the client bundle
A specific-conversation chat route (and others) crashed in the prod
container — `AsyncLocalStorage is not a constructor` — and every route
crashed under `vite dev`. Root cause: TanStack Start route files are
ISOMORPHIC, but they inlined `createServerFn` AND imported server-only
helpers at the top level. Those helper imports rode into the CLIENT
bundle; they crash when a helper transitively pulls a browser-throwing
node builtin (`node:async_hooks`, `node:crypto.createHash`). Prod
tree-shakes the side-effect-free ones, but `run-context`'s module-scope
`new AsyncLocalStorage()` defeats DCE and leaked to prod; dev tree-shakes
nothing, so it crashed broadly.

Fix follows the TanStack Start idiom (Intent skills start-core/
{execution-model,server-functions}):

1. Peel every inline `createServerFn` out of the 11 isomorphic routes
   into a sibling `*.functions.ts`. Routes now import server FUNCTIONS
   (RPC stubs, stripped from the client in dev AND prod), so the helpers
   never reach the browser.
2. Split the one mixed helper: `mappers.ts` keeps the pure naming
   helpers (client-safe); `createHash` content-hashing moves to a new
   server-only `select/source-content-hash.ts`.
3. Lazy-init `run-context`'s AsyncLocalStorage singleton so the module is
   import-side-effect-free and DCE drops it from every client chunk —
   this covers the non-route leak path (a client widget importing the
   `importSources` server fn's FILE, which chains to run-context).

No widget changes (components are isomorphic by design), no markers, no
`await import()` workarounds. API routes are unchanged — they're server
routes reached over HTTP, not imported, so they're already off the
client graph.

Verified on the prod container: 0 `AsyncLocalStorage` / 0 `createHash` in
`.output/public`; connect chat + governance + workflows + operating-model
+ reports all render with 0 console errors. tsc + biome clean; 1393/1393
unit tests pass.
