+++ packages/cockpit/src/routes/(app)/workspace/$wsId/operating-model.tsx
@@ -5,15 +5,10 @@
 
 import { Box, Center, Stack, Text } from "@mantine/core";
 import { ClientOnly, createFileRoute } from "@tanstack/react-router";
-import { createServerFn } from "@tanstack/react-start";
 
-import { loadOperatingModelGraph } from "#/tools/operating-model-load";
 import { ModelIcon } from "#/ui/cockpit/operating-model/nodes";
 import { OperatingModelCanvas } from "#/ui/cockpit/operating-model/operating-model-canvas";
-
-const loadModel = createServerFn({ method: "GET" }).handler(() =>
-	loadOperatingModelGraph(),
-);
+import { loadModel } from "./operating-model.functions";
 
 export const Route = createFileRoute("/(app)/workspace/$wsId/operating-model")({
 	loader: () => loadModel(),
