+++ packages/cockpit/src/routes/(app)/workspace/$wsId/reports/$reportId.tsx
@@ -14,7 +14,7 @@ import {
 	useNavigate,
 	useRouter,
 } from "@tanstack/react-router";
-import { createServerFn, useServerFn } from "@tanstack/react-start";
+import { useServerFn } from "@tanstack/react-start";
 import {
 	Check,
 	Pencil,
@@ -24,17 +24,14 @@ import {
 	X,
 } from "lucide-react";
 import { useState } from "react";
-import {
-	getReport,
-	renameReport,
-	setReportFingerprint,
-	softDeleteReport,
-	updateReportSummary,
-} from "#/db/cockpit/reports";
-import { computeReportFingerprint } from "#/duckdb/report-fingerprint-read";
-import { regenerateSummary } from "#/lib/report-summary-agent";
 import { ConfidenceStrip } from "#/ui/cockpit/widgets/answer-result";
 import { ResultGridWidget } from "#/ui/cockpit/widgets/result-grid";
+import {
+	deleteReportFn,
+	loadReport,
+	regenerateSummaryFn,
+	renameReportFn,
+} from "./$reportId.functions";
 
 // Report detail (DAT-624 / DAT-625) — the frozen artifact rendered over LIVE data:
 // the SQL is re-run on every open through the same result-grid stream, so numbers
@@ -47,56 +44,9 @@ import { ResultGridWidget } from "#/ui/cockpit/widgets/result-grid";
 // it's badged "outdated". A null stored fingerprint (pre-DAT-625 report, or a failed
 // mint-time fingerprint) is lazy-backfilled here — start tracking, show clean.
 //
-// The loader + action server fns are defined inline (the cockpit route convention)
-// so the plugin strips their cockpit_db + lake handlers from the client bundle.
-
-const loadReport = createServerFn({ method: "GET" })
-	.inputValidator((reportId: string) => reportId)
-	.handler(async ({ data: reportId }) => {
-		const report = await getReport(reportId);
-		if (!report) return null;
-		let outdated = false;
-		try {
-			const { fingerprint } = await computeReportFingerprint(report.sql);
-			if (report.summaryFingerprint === null) {
-				// First time we can fingerprint this report — backfill, don't badge.
-				await setReportFingerprint(report.id, fingerprint);
-			} else {
-				outdated = report.summaryFingerprint !== fingerprint;
-			}
-		} catch (err) {
-			// Best-effort: if the live result can't be fingerprinted (a since-broken
-			// SQL, a lake hiccup), don't badge — the grid surfaces the real error.
-			console.error("[reports] drift check failed — not flagging:", err);
-		}
-		return { report, outdated };
-	});
-
-const renameReportFn = createServerFn({ method: "POST" })
-	.inputValidator((data: { id: string; title: string }) => data)
-	.handler(async ({ data }) => {
-		await renameReport(data.id, data.title);
-	});
-
-const deleteReportFn = createServerFn({ method: "POST" })
-	.inputValidator((reportId: string) => reportId)
-	.handler(async ({ data: reportId }) => {
-		await softDeleteReport(reportId);
-	});
-
-// Regenerate (DAT-625): re-run the SQL, hand the old summary + fresh result to Haiku,
-// and persist the new summary together with the fresh fingerprint — the one path that
-// mutates `summary`. A throw here (missing report, LLM failure) propagates to the
-// caller, which keeps the old summary + outdated badge rather than half-applying.
-const regenerateSummaryFn = createServerFn({ method: "POST" })
-	.inputValidator((reportId: string) => reportId)
-	.handler(async ({ data: reportId }) => {
-		const report = await getReport(reportId);
-		if (!report) throw notFound();
-		const { fingerprint, result } = await computeReportFingerprint(report.sql);
-		const summary = await regenerateSummary(report.summary, result);
-		await updateReportSummary(reportId, summary, fingerprint);
-	});
+// The loader + action server fns live in the sibling `$reportId.functions.ts`
+// (the cockpit route convention) so their cockpit_db + lake handlers are stripped
+// from the client bundle.
 
 export const Route = createFileRoute(
 	"/(app)/workspace/$wsId/reports/$reportId",
